+++ src/api/controllers/stats/hourly/hourlyStatTransformer.spec.js
@@ -11,13 +11,25 @@ const expect = chai.expect;
 describe('hourly stat transformer', () => {
   it('should transform stat in response object', () => {
     const res = transformHourlyStat({
-      _id: 12,
+      _id: { hour: 12 },
       power: 12.4,
     });
 
     res.hour.should.be.eq(12);
     res.power.should.be.eq(12.4);
     res.deviceId.should.be.eq('');
+    Object.keys(res).length.should.be.eq(3);
+  });
+
+  it('should transform stat in response object with deviceId', () => {
+    const res = transformHourlyStat({
+      _id: { hour: 12, deviceId: '11:22:33:44' },
+      power: 12.4,
+    });
+
+    res.hour.should.be.eq(12);
+    res.power.should.be.eq(12.4);
+    res.deviceId.should.be.eq('11:22:33:44');
     Object.keys(res).length.should.be.eq(3);
   });
 });