+++ src/data/mongodb/stats/hourlyStatsProvider.js
@@ -63,13 +63,22 @@ export default function (database) {
       });
     },
 
-    getHourlyStat(dates, gateways) {
+    getHourlyStat(dates, gateways, groupFields) {
       const dayDates = dates.map(d => getRefDateTime(d));
 
       return new Promise((resolve, reject) => {
         db.collection(collectionName, (err, col) => {
           if (err) {
             reject(err);
+          }
+
+          let grouping = [];
+
+          if (groupFields) {
+            if (!Array.isArray(groupFields)) {
+              reject('Group field must be an array');
+            }
+            grouping = groupFields.map(field => ({ [field]: `$${field}` }));
           }
 
           col.aggregate([{
@@ -81,12 +90,13 @@ export default function (database) {
             },
           }, {
             $group: {
-              _id: { $hour: '$date' },
+              _id: Object.assign({}, { hour: { $hour: '$date' } }, ...grouping),
               power: {
                 $sum: '$power',
               },
             },
-          }])
+          },
+          ])
             .toArray((error, docs) => {
               if (error) {
                 reject(error);